+++ src/main/java/com/calendar/project/service/TagService.java
@@ -23,6 +23,6 @@ public interface TagService {
 
     public List<TagType> getTagsTypeList();
 
-    List<Tag> getTagsByEvent(Long EventId);
+    List<Tag> getTagsByEvent(int EventId);
 
 }
